+++ API/src/utils/email/templates/email_verification_template.ejs
@@ -123,7 +123,7 @@
                           </p>
                         </div>
                         <div class="contain">
-                          <a href="<%" ="verification_link" %>
+                          <a href=<%= verification_link %>
                             style="background:#009985;text-decoration:none
                             !important; font-weight:500; margin-top:35px;
                             color:#fff;text-transform:uppercase;
