--- API/src/utils/email/templates/tmp.html
@@ -1,183 +1,0 @@
-<!DOCTYPE html>
-<html lang="ar" dir="rtl">
-  <head>
-    <meta content="text/html; charset=utf-8" http-equiv="Content-Type" />
-    <title>قالب البريد الإلكتروني لإعادة تعيين كلمة المرور</title>
-    <meta
-      name="description"
-      content="قالب البريد الإلكتروني لإعادة تعيين كلمة المرور."
-    />
-    <style type="text/css">
-      a:hover {
-        text-decoration: underline !important;
-      }
-    </style>
-  </head>
-
-  <body
-    marginheight="0"
-    topmargin="0"
-    marginwidth="0"
-    style="margin: 0px; background-color: #f2f3f8"
-    leftmargin="0"
-  >
-    <table
-      cellspacing="0"
-      border="0"
-      cellpadding="0"
-      width="100%"
-      bgcolor="#f2f3f8"
-      style="
-        @import url(https://fonts.googleapis.com/css?family=Cairo:400,700&display=swap);
-        font-family: 'Cairo', sans-serif;
-      "
-    >
-      <tr>
-        <td>
-          <table
-            style="background-color: #f2f3f8; max-width: 670px; margin: 0 auto"
-            width="100%"
-            border="0"
-            align="center"
-            cellpadding="0"
-            cellspacing="0"
-          >
-            <tr>
-              <td style="height: 80px">&nbsp;</td>
-            </tr>
-            <tr>
-              <td style="text-align: center">
-                <a
-                  href="https://moocs-client.netlify.app/"
-                  title="logo"
-                  target="_blank"
-                >
-                  <img
-                    width="160"
-                    src="https://res.cloudinary.com/dipyrsqvy/image/upload/c_pad,b_auto:predominant,fl_preserve_transparency/v1680122215/moocs_yinmfd.jpg"
-                    title="logo"
-                    alt="logo"
-                  />
-                </a>
-              </td>
-            </tr>
-            <tr>
-              <td style="height: 20px">&nbsp;</td>
-            </tr>
-            <tr>
-              <td>
-                <table
-                  width="95%"
-                  border="0"
-                  align="center"
-                  cellpadding="0"
-                  cellspacing="0"
-                  style="
-                    max-width: 670px;
-                    background: #fff;
-                    border-radius: 3px;
-                    text-align: center;
-                    -webkit-box-shadow: 0 6px 18px 0 rgba(0, 0, 0, 0.06);
-                    -moz-box-shadow: 0 6px 18px 0 rgba(0, 0, 0, 0.06);
-                    box-shadow: 0 6px 18px 0 rgba(0, 0, 0, 0.06);
-                  "
-                >
-                  <tr>
-                    <td style="height: 40px">&nbsp;</td>
-                  </tr>
-                  <tr>
-                    <td style="padding: 0 35px">
-                      <div class="main-container">
-                        <div class="contain">
-                          <h1
-                            style="
-                              color: #1e1e2d;
-                              font-weight: 500;
-                              margin: 0;
-                              font-size: 32px;
-                              font-family: 'Rubik', sans-serif;
-                            "
-                          >
-                            تحقق من البريد الإلكتروني لـ Richie
-                          </h1>
-                        </div>
-                        <div class="contain">
-                          <span
-                            style="
-                              display: inline-block;
-                              vertical-align: middle;
-                              margin: 29px 0 26px;
-                              border-bottom: 1px solid #cecece;
-                              width: 100px;
-                            "
-                          ></span>
-                          <p
-                            style="
-                              color: #455056;
-                              font-size: 20px;
-                              line-height: 24px;
-                              margin: 0;
-                            "
-                          >
-                            مرحباً ${name}، مرحباً بك في منصة الابتكار المفتوحة!
-                            للمتابعة، يرجى تأكيد بريدك الإلكتروني عن طريق النقر
-                            على الرابط أدناه
-                          </p>
-                        </div>
-                        <div class="contain">
-                          <a
-                            href="${verification_link}"
-                            style="
-                              background: #009985;
-                              text-decoration: none !important;
-                              font-weight: 500;
-                              margin-top: 35px;
-                              color: #fff;
-                              text-transform: uppercase;
-                              font-size: 14px;
-                              padding: 10px 24px;
-                              display: inline-block;
-                              border-radius: 50px;
-                            "
-                          >
-                            تأكيد البريد الإلكتروني
-                          </a>
-                        </div>
-                      </div>
-                    </td>
-                  </tr>
-
-                  <tr>
-                    <td style="height: 40px">&nbsp;</td>
-                  </tr>
-                </table>
-              </td>
-            </tr>
-            <tr></tr>
-          </table>
-        </td>
-      </tr>
-      <tr>
-        <td style="height: 20px">&nbsp;</td>
-      </tr>
-      <tr>
-        <td style="text-align: center">
-          <p
-            style="
-              font-size: 14px;
-              color: #009985;
-              line-height: 18px;
-              margin: 0 0 0;
-            "
-          >
-            &copy;
-            <strong>https://moocs-client.netlify.app</strong>
-          </p>
-        </td>
-      </tr>
-      <tr>
-        <td style="height: 80px">&nbsp;</td>
-      </tr>
-    </table>
-  </body>
-</html>
